Centralise Claude provider state in useCodingAgentProviderState hook
- Add useCodingAgentProviderState hook that fetches providers/subscriptions
  internally, computes hasAnthropicProvider (all endpoint types, not just user),
  and auto-selects claude_code+api_key or claude_code+subscription via onChange
  once on first load, with a useRef guard to never override user selections
- Remove hasAnthropicProvider and hasClaudeSubscription props from CodingAgentForm;
  component now calls the hook itself
- Remove duplicated provider/subscription fetching, derivations, and auto-select
  effects from all five parent components (Onboarding, ProjectSettings,
  CreateProjectDialog, AgentSelectionModal, NewSpecTaskForm)
- Fix AppSettings hasAnthropicProvider to include global providers (was
  filtering to user-only endpoints)

Fixes: onboarding not auto-suggesting Claude Code when Anthropic is a global
provider, and showing "(not configured)" when a global Anthropic key exists.


Spec-Ref: helix-specs@c3df2aa0a:001544_issue-with-onboarding-on
